Fase Il/Backend/APIS Python/Clientes/requirements.txt
Flask
Flask-Cors
PyJWT
boto3
mysql-connector-python